+++ src/hooks/useUpdateUser.ts
@@ -1,10 +1,12 @@
 import useSWR from "swr";
+import { Preferences } from "../types";
 
 type UpdateUser = {
   id: string;
   name: string;
   mobile?: string | null;
   emergency?: string | null;
+  preferences?: Preferences | null;
 };
 
 export const updateUser = (user: UpdateUser) =>