+++ src/pages/ride/[id]/edit.tsx
@@ -9,6 +9,7 @@ import { getRide } from "../../api/ride";
 import { updateRide } from "../../../hooks";
 import { makeUtcDate, getFormRideDateAndTime } from "../../../../shared/utils";
 import { RideForm, FormValues } from "../../../components";
+import { Preferences } from "../../../types";
 
 type Props = {
   data: FormValues;
@@ -17,6 +18,7 @@ type Props = {
 const EditRide: NextPage<Props> = ({ data }: Props) => {
   const { data: session } = useSession();
   const user = session?.user;
+  const preferences = session?.preferences as Preferences;
   const { mutate } = useSWRConfig();
   const router = useRouter();
   const [waiting, setWaiting] = useState(false);
@@ -34,6 +36,7 @@ const EditRide: NextPage<Props> = ({ data }: Props) => {
   // Initial state for form: set name, leader and time
   const defaultValues = {
     ...data,
+    distance: parseInt(data.distance.toString(), 10),
     ...getFormRideDateAndTime(data.date),
   };
 
@@ -89,6 +92,7 @@ const EditRide: NextPage<Props> = ({ data }: Props) => {
           register={register}
           handleSubmit={handleSubmit(onSubmit)}
           waiting={waiting}
+          preferences={preferences}
         />
       </div>
     </>
@@ -100,6 +104,8 @@ export default EditRide;
 export const getServerSideProps: GetServerSideProps = async (context) => {
   const session = await getSession(context);
   const role = session && (session.role as string);
+  const preferences =
+    (session && (session.preferences as Preferences)) || undefined;
   const isAuthorised = !!session && role && ["LEADER", "ADMIN"].includes(role);
 
   if (!isAuthorised) {
@@ -112,7 +118,7 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
   }
 
   const { query } = context;
-  const data = await getRide(query.id, !!session);
+  const data = await getRide(query.id, preferences, !!session);
 
   return {
     props: {