+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2022-11-14
+
+Adds the first user preference in Settings menu; units in km or miles.
+
+### Added
+
+- Add default preferences (km) in constants and section in settings form
+- Convert and display all distances in correct units
+- Convert all submitted rides to use kms
+
+### Changed
+
+- Profile menu renamed to 'Settings'
+- Change update API and hook to include preferences
+
 ## [1.0.0] - 2022-11-14
 
 This is the first release of the rides app with essential features.